Misc changes to UI and FBX
Add a grid of the custom tool command button and recent commands button so that the two stay next to each other.

Remove FBX workaround code from the launcher since it's not needed anymore. Also move the code around so that it doesn't require extra booleans along with overrides in toolkits that have nothing to do with FBX.

Set up a binding for the auto FBX tooltip to say JMS instead of ASS in Halo 1 profiles
