+++ Launcher/ToolkitInterface/H1AToolkit.cs
@@ -70,7 +70,7 @@ namespace ToolkitLauncher.ToolkitInterface
         /// <returns></returns>
         public override async Task ImportModel(string path, ModelCompile importType, bool phantomFix, bool h2SelectionLogic, bool renderPRT, bool FPAnim, string characterFPPath, string weaponFPPath, bool accurateRender, bool verboseAnim, bool uncompressedAnim, bool skyRender, bool PDARender, bool resetCompression, bool autoFBX)
         {
-            if (autoFBX) { await AutoFBX.Model(this, path, importType, true); }
+            if (autoFBX) { await AutoFBX.Model(this, path, importType); }
 
             // todo(num0005): detect when the command is done running even when using -pause (and remove the forced shell usage)
             if (importType.HasFlag(ModelCompile.render))
